icon.library: $(PARANOIA_CFLAGS) cleanups

git-svn-id: https://svn.aros.org/svn/aros/trunk@37150 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/icon/bumprevision.c
@@ -96,7 +96,7 @@
 		snprintf (tempstr,
 		    sizeof (tempstr),
 		    "copy_%ld%s",
-		    copy_number,
+		    (long)copy_number,
 		    oldnameptr
 		);
 	    }
